+++ src/main/java/com/vnano/defenders/enchantment/ModEnchantments.java
@@ -5,6 +5,8 @@ public final class ModEnchantments {
     public static EnchantmentDefender FORTIFICATION;
     public static EnchantmentDefender REPRISAL;
     public static EnchantmentDefender FINESSE;
+    public static EnchantmentDefender REFLEXES;
+    public static EnchantmentDefender DEFLECTION;
 
     private ModEnchantments() {}
 }